% Generated by roxygen2 (4.0.1): do not edit by hand
\name{array2xml}
\alias{array2xml}
\title{This function provides a basic xml encoding used in the EVOp prototype wps services}
\usage{
array2xml(datetime, idata, filePath)
}
\arguments{
\item{datetime}{this is date&time format (often the index of a zoo time series object)}

\item{idata}{input array (often the core data of a zoo time series object)}

\item{filePath}{file path where file should be saved to}
}
\value{
saves xml file in the defined filePath
}
\description{
This function is used in wps processes based on R code to encode reults in xml format. It takes 3 inputs: datetime, idata and filePath. It transforms the input array 'idata' into an xml document based on a fixed tree. The datetime input is often the index of a zoo time series object, while the idata input is often the core data of a zoo time series object.
}
\author{
Claudia Vitolo
}
\references{
Environmental Virtual Observatory Pilot project: http://www.evo-uk.org/
}

